Hipify source in place
This is just the output of the hipify-inplace.sh and
hipify-tensoradapter.py scripts with no further modifications. I think
it's easier to review changes to the actual HIP source rather than
trying to thing about what the hipify script will do and since there are
a fair number of changes to review, that seems worth it. In the end we
should have a bunch of hip source files and .prehip cuda files that they
can be generated from. Then we can handle organization however we want:
restore the originals and have hipification be part of a build process,
have the hip versions on a separate branch, etc.

I'll check in the .prehip files in a separate commit to keep things a
bit cleaner.