Add ROCM to the DGL build
In my porting, I was finding it really annoying that everything in DGL
was hardcoded to the directory build/ and that it created build/
directories in various source subdirectories (which meant cleaning and
rebuilding was fraught). I modified things so that all sub-builds happen
in the main build directory.

There were also some bugs in the shell scripts and I cleaned them up a
bit to make them more robust.

Not all of this is strictly required for ROCM to build, so we might want
to strip it out. I already stripped out various warning silencing for
that reason.