Define warp_size as 64 for ROCM
This isn't really the right way for portable HIP code, which should instead query warp size at runtime (see https://rocm.docs.amd.com/projects/HIP/en/latest/how-to/hip_cpp_language_extensions.html#warpsize). But warp size is used as a compile-time constant in a number of places so that change is more invasive and for now we mostly care about gfx9*, so I'm punting on the right way.