Add ROCM to the DGL build (#3)
Adds CMake options to find ROCM (for a modern ROCM version) and configure it to build the source files with CUDA extensions.

In my porting, I was finding it really annoying that everything in DGL was hardcoded to the directory `build/`, as it made building with different configurations painful. The build also created `build/` directories in various source subdirectories (which meant cleaning and rebuilding was fraught). I modified things so that all sub-builds happen in the main build directory.

There were also some bugs in the sub-build shell scripts and I cleaned them up a bit to make them more robust.

Not all of this is strictly required for ROCM to build, so we might want to strip it out. I already stripped out the changes I made to silence various warnings for that reason.