Manual changes to DGL source for hipification (#10)
These changes are on the original DGL source code. These plus running `script/hipify-inplace.sh` yields the hipified version of DGL which is identical to the code currently in https://github.com/nod-ai/dgl/pull/9. The version in this commit should still run with CUDA.

This obviously shouldn't be merged into the same branch as PRs #1 through #9. The idea is that this would be the PR we would need for upstream (although I'm guessing they would actually want it in smaller chunks).