Define `warp_position` in GPU cache (#15)
If this isn't initialized, clang compiles away the if condition and
every thread increments the missing counter. I *think* this is a clang
bug because the shfl functions are supposed to have the values marked as
`maybe_undef` (which is literally what `maybe_undef` was created for),
but regardless initializing it is required here. This makes the gpu
cache tests pass when compiler optimizations are enabled.

Part of https://github.com/nod-ai/dgl/issues/14