Add a Dockerfile.rocm
This shows a docker image that can be used to compile and run DGL.

The rocprim version used is required for graphbolt. Unfortunately, this
leads to one test failing. TestRemainderPartition in tests/cpp/test_partition.cc.

Older versions of rocprim do not fail this test but do not support
graphbolt.