A script for creating hipified commits (#28)
This takes the current working tree and hipifies it, using the resulting working
tree to create a "fake" merge commit on the hipified branch (default
`hipify-inplace`) merging in the hipified changes but just taking the new
changes as given without actually trying to merge the contents.

The idea is that we can use this to populate the `hipify-inplace` convenience
branch (probably with a GitHub action eventually).