Redo the Docker images again
I'm back to operating in the rocm/pytorch image's sketchy conda
pseudo-environment. This appears to work, although pip yells at you
about running it as root.

By default, this builds the ROCm libraries and DGL with as many GPU
architectures as we can support, which unfortunately is not many due to
limitations in hip collections
(https://github.com/rocm/hipCollections/blob/b1cd36f600/CMakeLists.txt#L82).
Even with just these 4 the build for especially DGL is super slow and
due to some oddities in Ninja and DGL's recursive build setup, the
docker build appears to just hang while building graphbolt. I think this
is what we need for a turnkey build though.

I also swapped to just copying the context rather than cloning DGL from
a commit. This is annoying when based on the `hip-ready` branch, but I
couldn't find a less-bad option, as Docker per usual is *almost* good
but then actually falls on its face when you try to make it do something
useful. This also unfortunately makes it sort of necessary to copy the
whole .gitignore into the .dockerignore.