Avoid compiling GPU cache for invalid warp size
We hit static asserts if we do this, in addition to it extending compile
times. It turns out that RDNA with warp size 64 isn't even supported
with HIP also. I tried various methods and landed on `if constexpr`
comparing against the `warpSize` builtin. This is allegedly going to
become not constexpr, but it remains quite unclear to me what exactly is
going on. For instance `::rocm::device_warp_size()` may also become not
constexpr. Until there's some clarity, I'm going to use the established
`warpSize` builtin rather than adding an additional macro that may be no
better.

I also preserved the ability to use a warp size smaller than the device
native warp size. The README and some checks in the code imply that this
is allowed. I haven't actually seen that work and in the original codes
templates with a different warp size aren't even instantiated (and
therefore fail to link), but leaving that functionality in place just in
case.

I ended up not running clang-format because it did a lot of extraneous
reformatting. I think keeping the diff here with upstream small is a
higher priority than strict line length limit adherence.